LayoutTests/media/modern-media-controls/playback-support/playback-support-button-click-expected.txt

Testing the PlaybackSupport behavior by clicking on the play/pause button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media is playing
PASS mediaController.controls.playPauseButton.playing is true

Pressing on the play/pause button
Media is paused
PASS mediaController.controls.playPauseButton.playing is false

Pressing on the play/pause button
Media is playing
PASS mediaController.controls.playPauseButton.playing is true

PASS successfullyParsed is true

TEST COMPLETE
